backend/requirements.txt
fastapi
uvicorn
pydantic
motor
passlib
python-dotenv